app_python/requirements.txt
# Production
Flask==3.1.0
python-dotenv==1.0.1
Werkzeug==3.1.3

# Development & Testing
pytest==8.2.2
pytest-cov==5.0.0
flake8==7.1.1